+++ src/components/common-part/DisplayProductProps.ts
